backend/src/module/quant/extras/requirements.txt
# Excel Agent Core
openpyxl>=3.1.0
pydantic==2.11.7
pydantic-ai==1.11.0
pydantic-evals==1.11.0
fastmcp==2.12.4
openai>=1.0.0
rich>=13.0.0

# Demo: ExcelAssistant (FastAPI web app)
fastapi>=0.100.0
uvicorn>=0.20.0

# Demo: SlackExcelWorkflow (Slack bot)
python-dotenv>=1.0.0
requests>=2.28.0
slack-bolt>=1.18.0
